cli/pyproject.toml
[project]
name = "bracc-cli"
version = "0.1.0"
description = "BR-ACC CLI for local operations, API queries, and ETL workflows"
requires-python = ">=3.12"
license = "AGPL-3.0-or-later"
dependencies = [
    "httpx>=0.28.0",
    "python-dotenv>=1.0.1",
    "rich>=13.9.4",
    "typer>=0.16.1",
]

[project.scripts]
bracc = "bracc_cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["bracc_cli"]

[dependency-groups]
dev = [
    "ruff>=0.15.2",
]
